requirements.txt
google-generativeai
pillow
json
dotenv
